GUI - Modify add API proxy to nginx server
This add a new location in the nginx server with proxy_pass
to make calls to the backend service.

For now I'm leaving the upstream definition which I was using
with docker on my local machine.

`host.docker.internal` is a special DNS within the Docker which is
supposed to be used **for development only**.

Quoting from the Docker documentation:
> From 18.03 onwards our recommendation is to connect to the special
> DNS name host.docker.internal, which resolves to the internal
> IP address used by the host.
> This is for development purpose and will not work in a production environment
> outside of Docker Desktop for Mac.