GUI - Configure the development proxy for backend api calls
To make valid requests to the backend we need to add X-Simple-Paste header.
The create-react-app 'proxy' entry in the package.json is not enough to
do it. This is why a proxy server is configured using
`http-proxy-middleware`.

It is also necessary to cut the `/api` fragment from url. ✂️

See create-react-app docs for more info:
https://create-react-app.dev/docs/proxying-api-requests-in-development/